Moving setup.new in init.rs
